+++ client/src/features/interview-question/index.ts
@@ -1,1 +1,2 @@
-export { default as ExpectedQuestionLayout } from './ExpectedQuestionLayout';
+export { default as UserQuestionPlayList } from './playlist/UserQuestionPlayList';
+export { default as ExpectedQuestionSelector } from './expected-question/ExpectedQuestionSelector';
