+++ client/src/widgets/interview/ExpectedQuestionLayout.tsx
@@ -3,8 +3,7 @@ import styled from 'styled-components';
 import { ExplanationContent } from '@/src/shared/ui';
 import { styleMixin, V } from '@/src/shared/styles';
 
-import { UserQuestionPlayList } from './playlist';
-import { ExpectedQuestionSelector } from './expected-question';
+import { UserQuestionPlayList, ExpectedQuestionSelector } from '@/src/features/interview-question';
 
 const explanationList = [
   '면접 질문 셋팅 가이드',
