+++ client/src/widgets/interview/index.ts
@@ -1,0 +1,1 @@
+export { default as ExpectedQuestionLayout } from './ExpectedQuestionLayout';
